Pulled generateKeys() up to DockerBaseIT because more than Qpid tests will generate keys
